# docs/API.yaml
openapi: 3.1.0
info:
  title: Wyvern Drive API
  description: |
    Internal API documentation for Wyvern Drive Supabase Edge Function.

    **⚠️ INTERNAL USE ONLY - NOT FOR PUBLIC DISTRIBUTION**

    This API handles file metadata storage, versioning, sharing, and video streaming.
    Actual file data is stored on Discord CDN; this API manages the index.
  version: 1.0.0
  contact:
    name: Developer
servers:
  - url: https://{project-ref}.supabase.co/functions/v1/api
    description: Production (Supabase Edge Function)
    variables:
      project-ref:
        default: your-project-ref
...
paths:
...
  /files/{userId}:
    get:
      tags: [Files]
      summary: Get file tree
      description: Returns all files and folders for a user as a nested tree structure.
      operationId: getFiles
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: File tree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileTree'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Files]
      summary: Create file or directory
      description: |
        Creates a new file or folder. If a file with the same name exists,
        the existing content is versioned and the new content replaces it.
      operationId: createFile
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                parent_id:
                  type: integer
                  nullable: true
                  description: Parent folder ID (null for root)
                name:
                  type: string
                type:
                  type: string
                  enum: [file, directory]
                size:
                  type: integer
                  description: File size in bytes
                content:
                  type: string
                  description: JSON array of chunk info
                encrypted:
                  type: boolean
                encryption_salt:
                  type: string
                  description: Base64 salt for encrypted files
      responses:
        '200':
          description: Created file ID
          content:
            application/json:
              schema:
                type: integer
        '500':
          $ref: '#/components/responses/ServerError'
  /files/{userId}/{fileId}:
    delete:
      tags: [Files]
      summary: Delete file or empty directory
      operationId: deleteFile
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/fileId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          description: Directory not empty
        '500':
          $ref: '#/components/responses/ServerError'
  /files/{userId}/{fileId}/update:
    post:
      tags: [Files]
      summary: Update file metadata
      description: Update file properties (name, parent_id, etc.)
      operationId: updateFile
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/fileId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                parent_id:
                  type: integer
                  nullable: true
      responses:
        '200':
          $ref: '#/components/responses/Success'
  /files/{userId}/{fileId}/recursive:
    delete:
      tags: [Files]
      summary: Delete directory recursively
      description: Deletes a folder and all its contents.
      operationId: deleteRecursive
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/fileId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
...
  /versions/{userId}/{fileId}:
    get:
      tags: [Versions]
      summary: Get version history
      operationId: getVersions
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/fileId'
      responses:
        '200':
          description: List of versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Version'

    post:
      tags: [Versions]
      summary: Create version manually
      operationId: createVersion
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/fileId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                size:
                  type: integer
      responses:
        '200':
          description: Created version
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  versionNumber:
                    type: integer
  /versions/{userId}/{fileId}/restore/{versionId}:
    post:
      tags: [Versions]
      summary: Restore file to version
      operationId: restoreVersion
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/fileId'
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '404':
          description: Version not found
  /versions/{userId}/{fileId}/{versionId}:
    delete:
      tags: [Versions]
      summary: Delete version
      operationId: deleteVersion
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/fileId'
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/Success'
...
components:
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: User identifier (hashed webhook URL)
      schema:
        type: string
    fileId:
      name: fileId
      in: path
      required: true
      schema:
        type: integer
  schemas:
    FileTree:
      type: object
      properties:
        children:
          type: object
          additionalProperties:
            oneOf:
              - $ref: '#/components/schemas/FileEntry'
              - $ref: '#/components/schemas/FolderEntry'
    FileEntry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [file]
        size:
          type: integer
        content:
          type: string
          description: JSON chunk array
        encrypted:
          type: integer
          enum: [0, 1]
        encryption_salt:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FolderEntry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [directory]
        children:
          type: object
          additionalProperties: true
    Version:
      type: object
      properties:
        id:
          type: integer
        version_number:
          type: integer
        size:
          type: integer
        created_at:
          type: string
          format: date-time
...
  responses:
    Success:
      description: Operation successful
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string